meter update #2
Took me a (((bit))) of time since I'm focusing on other things, but the second meter update is here :)
This update includes all of the item effect meters for the classes I haven't gotten into in the first one.
You don't need to go into much depth examining the files to see how inconsistent they are on pretty much every level, but as far as I've checked everything apart from the phlog works. Not like I'm doing a disservice to anyone ¯\_(シ)_/¯